Back to original

+++ jdk/src/share/native/sun/java2d/loops/GraphicsPrimitiveMgr.c
@@ -320,7 +320,7 @@ static jboolean InitSimpleTypes
     jfieldID field;
     jobject obj;
 
-    for (pHdr = pStart; ((long)pHdr) < ((long)pEnd); pHdr = PtrAddBytes(pHdr, size)) {
+    for (pHdr = pStart; pHdr < pEnd; pHdr = PtrAddBytes(pHdr, size)) {
         field = (*env)->GetStaticFieldID(env,
                                          SimpleClass,
                                          pHdr->Name,
